+++ airflow/decorators/setup_teardown.py
@@ -22,34 +22,27 @@ from typing import Callable
 from airflow import AirflowException
 from airflow.decorators import python_task
 from airflow.decorators.task_group import _TaskGroupFactory
-from airflow.settings import _ENABLE_AIP_52
 
 
 def setup_task(func: Callable) -> Callable:
-    if not _ENABLE_AIP_52:
-        raise AirflowException("AIP-52 Setup tasks are disabled.")
-
     # Using FunctionType here since _TaskDecorator is also a callable
     if isinstance(func, types.FunctionType):
         func = python_task(func)
     if isinstance(func, _TaskGroupFactory):
         raise AirflowException("Task groups cannot be marked as setup or teardown.")
-    func._is_setup = True  # type: ignore[attr-defined]
+    func.is_setup = True  # type: ignore[attr-defined]
     return func
 
 
 def teardown_task(_func=None, *, on_failure_fail_dagrun: bool = False) -> Callable:
-    if not _ENABLE_AIP_52:
-        raise AirflowException("AIP-52 Teardown tasks are disabled.")
-
     def teardown(func: Callable) -> Callable:
         # Using FunctionType here since _TaskDecorator is also a callable
         if isinstance(func, types.FunctionType):
             func = python_task(func)
         if isinstance(func, _TaskGroupFactory):
             raise AirflowException("Task groups cannot be marked as setup or teardown.")
-        func._is_teardown = True  # type: ignore[attr-defined]
-        func._on_failure_fail_dagrun = on_failure_fail_dagrun  # type: ignore[attr-defined]
+        func.is_teardown = True  # type: ignore[attr-defined]
+        func.on_failure_fail_dagrun = on_failure_fail_dagrun  # type: ignore[attr-defined]
         return func
 
     if _func is None: