+++ airflow/providers/ssh/hooks/ssh.py
@@ -31,7 +31,7 @@ from sshtunnel import SSHTunnelForwarder
 from tenacity import Retrying, stop_after_attempt, wait_fixed, wait_random
 
 from airflow.compat.functools import cached_property
-from airflow.exceptions import AirflowException
+from airflow.exceptions import AirflowException, AirflowProviderDeprecationWarning
 from airflow.hooks.base import BaseHook
 from airflow.utils.platform import getuser
 from airflow.utils.types import NOTSET, ArgNotSet
@@ -172,7 +172,7 @@ class SSHHook(BaseHook):
                         "Extra option `timeout` is deprecated."
                         "Please use `conn_timeout` instead."
                         "The old option `timeout` will be removed in a future version.",
-                        DeprecationWarning,
+                        AirflowProviderDeprecationWarning,
                         stacklevel=2,
                     )
                     self.timeout = int(extra_options["timeout"])
@@ -232,7 +232,7 @@ class SSHHook(BaseHook):
                 "Parameter `timeout` is deprecated."
                 "Please use `conn_timeout` instead."
                 "The old option `timeout` will be removed in a future version.",
-                DeprecationWarning,
+                AirflowProviderDeprecationWarning,
                 stacklevel=1,
             )
 
@@ -368,7 +368,7 @@ class SSHHook(BaseHook):
             "The contextmanager of SSHHook is deprecated."
             "Please use get_conn() as a contextmanager instead."
             "This method will be removed in Airflow 2.0",
-            category=DeprecationWarning,
+            category=AirflowProviderDeprecationWarning,
         )
         return self
 
@@ -434,7 +434,7 @@ class SSHHook(BaseHook):
             "use get_tunnel() instead. But please note that the"
             "order of the parameters have changed"
             "This method will be removed in Airflow 2.0",
-            category=DeprecationWarning,
+            category=AirflowProviderDeprecationWarning,
         )
 
         return self.get_tunnel(remote_port, remote_host, local_port)